# Integration PR followups: make_workdiv, uniform_elements, concrete kernel dimensions