--- a/HeterogeneousCore/AlpakaInterface/interface/alpakastdAlgorithm.h
+++ b/HeterogeneousCore/AlpakaInterface/interface/alpakastdAlgorithm.h
@@ -7,11 +7,27 @@
 
 #include <alpaka/alpaka.hpp>
 
-// reimplementation of std algorithms able to compile with Alpaka,
-// mostly by declaring them constexpr (until C++20, which will make it
-// constexpr by default. TODO: drop when moving to C++20)
+// reimplementation of std algorithms able to work on device code
 
 namespace alpaka_std {
+
+  template <typename RandomIt, typename T, typename Compare = std::less<T>>
+  ALPAKA_FN_HOST_ACC constexpr RandomIt lower_bound(RandomIt first, RandomIt last, const T &value, Compare comp = {}) {
+    auto count = last - first;
+
+    while (count > 0) {
+      auto it = first;
+      auto step = count / 2;
+      it += step;
+      if (comp(*it, value)) {
+        first = ++it;
+        count -= step + 1;
+      } else {
+        count = step;
+      }
+    }
+    return first;
+  }
 
   template <typename RandomIt, typename T, typename Compare = std::less<T>>
   ALPAKA_FN_HOST_ACC constexpr RandomIt upper_bound(RandomIt first, RandomIt last, const T &value, Compare comp = {}) {