I started to address some follow-up tasks in #75. In particular:

- Switch to `cms::alpakatools::makeworkdir` instead of our custom `createWorkDiv`.
- Switch to `cms::alpakatools::uniform_elements` for kernel loops.
- Don't set a custom work division for CPU.
- Started moving towards the removal of `kVerticalModuleSlope`.
- Use concrete kernel dimensions instead of templated ones.