@ariostas 2024-12-17
I think the plots might look a bit different now that the work division is different, but hopefully they are still run-to-run reproducible.

For now, I'm just making sure I didn't break something.

/run standalone
/run checks