@ariostas 2024-12-17
Hopefully it's fine now.

/run standalone