@ariostas 2024-12-18
Sorry for all the force-pushing.

I made the code compatible with both `kVerticalModuleSlope` and infinity, so that we can change the data files without breaking anything. Once that is done we can fully remove it.

I think the last low hanging fruit that I'll include here is to set concrete dimensions instead of templated types for kernels and alpaka functions.