@slava77 2024-12-20
/run all