@ariostas 2025-01-10
/run all