@slava77 2025-01-10
before the break we talked about batching, assuming other contributions overlapping this PR can accumulate.
It looks like there is nothing ready yet. So, perhaps it's more practical now to just submit this upstream to cms-sw/cmssw already.
Commits are linear and few, I don't see a need to squash or rebase.