@slava77 2025-01-10
this crashed in the bot tests with GPU: did it run locally on GPU or is it some more recent feature in the IBs?