@ariostas 2025-01-13
Seems like the bug is outside of the changes I made for this PR. The issue was triggered by the switch from `int8_t` to `int16_t`. My guess is that something was left unset and for `0xff` it was fine, but `0xffff` is large enough to cause a segfault. I'm still looking into where this is coming from.