@ariostas 2025-01-13
Nevermind, seems like it is a problem in this PR. Maybe I'm misunderstanding how some other functionality works on GPUs.