@ariostas 2025-01-13
Turns out that https://github.com/cms-sw/cmssw/pull/46967 broke running on GPUs. The binary search is not working for some reason.