@VourMa 2025-01-13
> Turns out that [cms-sw#46967](https://github.com/cms-sw/cmssw/pull/46967) broke running on GPUs. The binary search is not working for some reason.

The GPU tests passed back then though:
https://github.com/cms-sw/cmssw/pull/46967#issuecomment-2549160037
What changed?