@slava77 2025-01-13
> The GPU tests passed back then though:
> [cms-sw#46967 (comment)](https://github.com/cms-sw/cmssw/pull/46967#issuecomment-2549160037)
> What changed?

the [baseline](https://github.com/cms-sw/cmssw/pull/47084#issuecomment-2584483650) in the PR tests ran OK as well. So, the crash appears from just the incremental addition of this PR.
Well, unless there is a non-reproducible component and it may or may not crash unpredictably.