@ariostas 2025-01-14
> the baseline in the PR tests ran OK as well.

The baseline doesn't crash, but it produces garbage results. With the switch from `int8_t` to `int16_t` in this PR the garbage is bad enough to cause a segfault.